pyproject.toml
[project]
name = "boostspace-cli"
version = "0.1.0"
description = "CLI to manage Boost.space / Make.com workflows from the terminal"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "keyring>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
boost = "boostspace_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/boostspace_cli"]

[tool.hatch.build.targets.wheel.force-include]
"src/boostspace_cli/data/module_registry.json" = "boostspace_cli/data/module_registry.json"
"src/boostspace_cli/data/native_modules.json" = "boostspace_cli/data/native_modules.json"
"src/boostspace_cli/data/formula_registry.json" = "boostspace_cli/data/formula_registry.json"

[tool.hatch.build.targets.sdist]
include = [
  "src/boostspace_cli/data/module_registry.json",
  "src/boostspace_cli/data/native_modules.json",
  "src/boostspace_cli/data/formula_registry.json",
]
